separating schema into separate file
